Rename TopologyControls to TopologySidePanel
The only control in the component was side panel related.
