+++ src/components/topology/TopologySidePanel.vue
